+++ package.json
@@ -4,7 +4,7 @@
   "main": "index.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "serve": "ts-node --files ./src/app.ts"
+    "serve": "ts-node --files ./src/server.ts"
   },
   "keywords": [],
   "author": "",
